Update BlackflyCamera.py
Added a commented out, updated version of the calculate_statistics function that re-calculates the centroid of the image taking into account the first centroid. This code assumes a size of the beam on the camera so if it is to be used in a general sense, the parameters of this cut may have to modified.